Combine utils.py ndarray.py to fallback.py
