+++ cupyx/fallback_mode/fallback.py
@@ -1,15 +1,18 @@
 """
-Main fallback class.
+`fallback_mode` for cupy. Whenever a method is not yet implemented in CuPy,
+it will fallback to corresponding NumPy method.
 """
-
+import sys
 import types
+
 import numpy as np
 
 import cupy as cp
+from cupyx.fallback_mode import data_transfer
 
-
-from cupyx.fallback_mode import utils
-from cupyx.fallback_mode import ndarray
+# -----------------------------------------------------------------------------
+# _RecursiveAttr
+# -----------------------------------------------------------------------------
 
 
 class _RecursiveAttr:
@@ -32,7 +35,7 @@ class _RecursiveAttr:
         separately.
         """
         if self._cupy_object is cp.ndarray:
-            return isinstance(instance, ndarray.ndarray)
+            return isinstance(instance, ndarray)
 
         if self._cupy_object is not None:
             return isinstance(instance, self._cupy_object)
@@ -104,10 +107,188 @@ class _RecursiveAttr:
 
         # Execute cupy method
         if self._cupy_object is not None:
-            return utils._call_cupy(self._cupy_object, args, kwargs)
+            return _call_cupy(self._cupy_object, args, kwargs)
 
         # Execute numpy method
-        return utils._call_numpy(self._numpy_object, args, kwargs)
+        return _call_numpy(self._numpy_object, args, kwargs)
 
 
 numpy = _RecursiveAttr(np, cp)
+
+# -----------------------------------------------------------------------------
+# utils
+# -----------------------------------------------------------------------------
+
+
+def _call_cupy(func, args, kwargs):
+    """
+    Calls cupy function with *args and **kwargs and
+    does necessary data transfers.
+
+    Args:
+        func: A cupy function that needs to be called.
+        args (tuple): Arguments.
+        kwargs (dict): Keyword arguments.
+
+    Returns:
+        Result after calling func and performing data transfers.
+    """
+
+    args, kwargs = _get_cupy_ndarray(args, kwargs)
+
+    res = func(*args, **kwargs)
+
+    return _get_fallback_ndarray(res)
+
+
+def _call_numpy(func, args, kwargs):
+    """
+    Calls numpy function with *args and **kwargs and
+    does necessary data transfers.
+
+    Args:
+        func: A numpy function that needs to be called.
+        args (tuple): Arguments.
+        kwargs (dict): Keyword arguments.
+
+    Returns:
+        Result after calling func and performing data transfers.
+    """
+
+    args, kwargs = _get_cupy_ndarray(args, kwargs)
+
+    numpy_args, numpy_kwargs = data_transfer._get_numpy_args(args, kwargs)
+
+    numpy_res = func(*numpy_args, **numpy_kwargs)
+
+    cupy_res = data_transfer._get_cupy_result(numpy_res)
+
+    return _get_fallback_ndarray(cupy_res)
+
+
+# -----------------------------------------------------------------------------
+# ndarray wrapper and proxy magic methods
+# -----------------------------------------------------------------------------
+
+
+class ndarray:
+    """
+    Wrapper around cupy.ndarray
+    Gets initialized with a cupy ndarray.
+    """
+
+    def __init__(self, array):
+        self._array = array
+
+    def __getattr__(self, attr):
+        """
+        Catches attributes corresponding to ndarray.
+
+        Args:
+            attr (str): Attribute of ndarray class.
+
+        Returns:
+            (_RecursiveAttr object, self._array.attr):
+            Returns_RecursiveAttr object with numpy_object, cupy_object.
+            Returns self._array.attr if attr is not callable.
+        """
+
+        cupy_object = getattr(cp.ndarray, attr, None)
+
+        numpy_object = getattr(np.ndarray, attr)
+
+        if not callable(numpy_object):
+            return getattr(self._array, attr)
+
+        return _RecursiveAttr(numpy_object, cupy_object, self)
+
+    def _get_array(self):
+        """
+        Returns _array (cupy.ndarray) of ndarray object.
+        """
+        return self._array
+
+
+def _get_cupy_ndarray(args, kwargs):
+    return data_transfer._get_xp_args(
+        ndarray, ndarray._get_array, (args, kwargs))
+
+
+def _get_fallback_ndarray(cupy_res):
+    return data_transfer._get_xp_args(cp.ndarray, ndarray, cupy_res)
+
+
+# Decorator for ndarray magic methods
+def make_method(name):
+    def method(self, *args, **kwargs):
+
+        args, kwargs = _get_cupy_ndarray(args, kwargs)
+
+        cupy_method = getattr(cp.ndarray, name)
+
+        res = cupy_method(self._array, *args, **kwargs)
+
+        return _get_fallback_ndarray(res)
+
+    return method
+
+
+def _create_magic_methods():
+    """
+    Set magic methods of cupy.ndarray as methods of utils.ndarray.
+    """
+
+    _common = [
+
+        # Comparison operators:
+        '__eq__', '__ne__', '__lt__', '__gt__', '__le__', '__ge__',
+
+        # Unary operations:
+        '__neg__', '__pos__', '__abs__', '__invert__',
+
+        # Arithmetic:
+        '__add__', '__sub__', '__mul__', '__truediv__', '__floordiv__',
+        '__mod__', '__divmod__', '__pow__', '__lshift__', '__rshift__',
+        '__and__', '__or__', '__xor__',
+
+        # Arithmetic, in-place:
+        '__iadd__', '__isub__', '__imul__', '__itruediv__', '__ifloordiv__',
+        '__imod__', '__ipow__', '__ilshift__', '__irshift__',
+        '__iand__', '__ior__', '__ixor__',
+
+        # reflected-methods:
+        '__radd__', '__rsub__', '__rmul__', '__rtruediv__', '__rfloordiv__',
+        '__rmod__', '__rdivmod__', '__rpow__', '__rlshift__', '__rrshift__',
+        '__rand__', '__ror__', '__rxor__',
+
+        # For standard library functions:
+        '__copy__', '__deepcopy__', '__reduce__',
+
+        # Container customization:
+        '__iter__', '__len__', '__getitem__', '__setitem__',
+
+        # Conversion:
+        '__int__', '__float__', '__complex__',
+
+        # String representations:
+        '__repr__', '__str__'
+    ]
+
+    _py3 = [
+        '__matmul__', '__rmatmul__', '__bool__'
+    ]
+
+    _py2 = [
+        '__div__', '__rdiv__', '__idiv__', '__nonzero__',
+        '__long__', '__hex__', '__oct__'
+    ]
+
+    _specific = _py3
+    if sys.version_info[0] == 2:
+        _specific = _py2
+
+    for method in _common + _specific:
+        setattr(ndarray, method, make_method(method))
+
+
+_create_magic_methods()
